TACHYON-799: Organize the classes into proper packages under tachyon security
