+++ common/src/test/java/tachyon/security/login/LoginModuleTest.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.security;
+package tachyon.security.login;
 
 import java.security.Principal;
 
@@ -23,9 +23,13 @@ import javax.security.auth.login.LoginContext;
 import org.junit.Assert;
 import org.junit.Test;
 
+import tachyon.security.User;
+import tachyon.security.login.TachyonJaasConfiguration;
+import tachyon.security.login.TachyonJaasProperties;
+
 /**
- * Unit test for the login modules defined in {@link tachyon.security.TachyonLoginModule} and
- * used in {@link tachyon.security.TachyonJaasConfiguration}
+ * Unit test for the login modules defined in {@link tachyon.security.login.TachyonLoginModule} and
+ * used in {@link tachyon.security.login.TachyonJaasConfiguration}
  */
 public class LoginModuleTest {
